// apps/frontend/packages/api/base/codegen/types/FactForm2.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { FactFormCode2 } from "./FactFormCode2";
import type { FactFormVersion2 } from "./FactFormVersion2";

export interface FactForm2 {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type integer
    */
    factPackageId: number;
    /**
     * @type string
    */
    code: FactFormCode2;
    /**
     * @type string
    */
    label: string;
    /**
     * @type boolean
    */
    isFilled: boolean;
    currentVersion?: (FactFormVersion2 | null);
    /**
     * @description Полная история версий формы, отсортирована по возрастанию versionNumber
     * @type array | undefined
    */
    versions?: FactFormVersion2[];
}