// apps/frontend/packages/api/base/codegen/types/FactPackageRemarkCreateInput.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export interface FactPackageRemarkCreateInput {
    /**
     * @type integer
    */
    relatedFormId: number;
    /**
     * @type string
    */
    description: string;
    /**
     * @type string
    */
    requiredAction: string;
}